migrate generete ground truth
