+++ src/spikeinterface/benchmark/tests/test_benchmark_sorter_without_gt.py
@@ -2,7 +2,7 @@ import shutil
 import pytest
 from pathlib import Path
 
-from spikeinterface import generate_ground_truth_recording
+from spikeinterface.generation import generate_ground_truth_recording
 from spikeinterface.preprocessing import bandpass_filter
 from spikeinterface.benchmark import SorterStudyWithoutGroundTruth
