+++ src/spikeinterface/core/tests/test_sorting_tools.py
@@ -4,7 +4,7 @@ import numpy as np
 
 from spikeinterface.core import NumpySorting
 
-from spikeinterface.core import generate_ground_truth_recording
+from spikeinterface.generation import generate_ground_truth_recording
 from spikeinterface.core.sorting_tools import (
     spike_vector_to_spike_trains,
     random_spikes_selection,
@@ -52,7 +52,7 @@ def test_random_spikes_selection():
         num_channels=10,
         num_units=5,
         generate_sorting_kwargs=dict(firing_rates=10.0, refractory_period_ms=4.0),
-        noise_kwargs=dict(noise_levels=5.0, strategy="tile_pregenerated"),
+        noise_kwargs=dict(noise_levels=5.0),
         seed=2205,
     )
     max_spikes_per_unit = 12
