+++ src/spikeinterface/core/tests/test_template_tools.py
@@ -1,7 +1,8 @@
 import pytest
 import numpy as np
 
-from spikeinterface.core import generate_ground_truth_recording, create_sorting_analyzer
+from spikeinterface.core import create_sorting_analyzer
+from spikeinterface.generation import generate_ground_truth_recording
 
 
 from spikeinterface import Templates
@@ -19,7 +20,7 @@ def get_sorting_analyzer():
         sampling_frequency=10_000.0,
         num_channels=4,
         num_units=10,
-        noise_kwargs=dict(noise_levels=5.0, strategy="tile_pregenerated"),
+        noise_kwargs=dict(noise_levels=5.0),
         seed=2205,
     )
     recording.annotate(is_filtered=True)
