+++ src/spikeinterface/core/tests/test_time_series_tools.py
@@ -25,7 +25,6 @@ def test_write_binary(tmp_path):
         durations=durations,
         num_channels=num_channels,
         sampling_frequency=sampling_frequency,
-        strategy="tile_pregenerated",
     )
     file_paths = [tmp_path / "binary01.raw"]
 
@@ -52,7 +51,6 @@ def test_write_binary_offset(tmp_path):
         durations=durations,
         num_channels=num_channels,
         sampling_frequency=sampling_frequency,
-        strategy="tile_pregenerated",
     )
     file_paths = [tmp_path / "binary01.raw"]
 
@@ -85,7 +83,6 @@ def test_write_binary_parallel(tmp_path):
         num_channels=num_channels,
         sampling_frequency=sampling_frequency,
         dtype=dtype,
-        strategy="tile_pregenerated",
     )
     file_paths = [tmp_path / "binary01.raw", tmp_path / "binary02.raw"]
 
@@ -115,7 +112,6 @@ def test_write_binary_multiple_segment(tmp_path):
         durations=durations,
         num_channels=num_channels,
         sampling_frequency=sampling_frequency,
-        strategy="tile_pregenerated",
     )
     file_paths = [tmp_path / "binary01.raw", tmp_path / "binary02.raw"]
 
@@ -136,9 +132,7 @@ def test_write_binary_multiple_segment(tmp_path):
 
 def test_write_memory_recording():
     # 2 segments
-    recording = MockRecording(
-        num_channels=2, durations=[10.325, 3.5], sampling_frequency=30_000, strategy="tile_pregenerated"
-    )
+    recording = MockRecording(num_channels=2, durations=[10.325, 3.5], sampling_frequency=30_000)
     recording = recording.save()
 
     # write with loop
