+++ src/spikeinterface/core/tests/test_waveforms_extractor_backwards_compatibility.py
@@ -6,7 +6,8 @@ import shutil
 
 import numpy as np
 
-from spikeinterface.core import generate_ground_truth_recording, SortingAnalyzer
+from spikeinterface.core import SortingAnalyzer
+from spikeinterface.generation import generate_ground_truth_recording
 from spikeinterface.core.core_tools import SIJsonEncoder
 
 from spikeinterface.core.waveforms_extractor_backwards_compatibility import MockWaveformExtractor
@@ -34,7 +35,7 @@ def get_dataset():
                 alpha=(100.0, 500.0),
             )
         ),
-        noise_kwargs=dict(noise_levels=5.0, strategy="tile_pregenerated"),
+        noise_kwargs=dict(noise_levels=5.0),
         seed=2406,
     )
     return recording, sorting
