+++ src/spikeinterface/exporters/tests/common.py
@@ -1,7 +1,8 @@
 import pytest
 from pathlib import Path
 
-from spikeinterface.core import generate_ground_truth_recording, create_sorting_analyzer, compute_sparsity
+from spikeinterface.core import create_sorting_analyzer, compute_sparsity
+from spikeinterface.generation import generate_ground_truth_recording
 
 
 def make_sorting_analyzer(sparse=True, with_group=False):
@@ -18,7 +19,7 @@ def make_sorting_analyzer(sparse=True, with_group=False):
             contact_shape_params={"radius": 6},
         ),
         generate_sorting_kwargs=dict(firing_rates=10.0, refractory_period_ms=4.0),
-        noise_kwargs=dict(noise_levels=5.0, strategy="on_the_fly"),
+        noise_kwargs=dict(noise_levels=5.0),
         seed=2205,
     )
