+++ src/spikeinterface/extractors/tests/test_mdaextractors.py
@@ -1,7 +1,7 @@
 import pytest
 from pathlib import Path
 from spikeinterface.core.testing import check_recordings_equal, check_sortings_equal
-from spikeinterface.core import generate_ground_truth_recording
+from spikeinterface.generation import generate_ground_truth_recording
 from spikeinterface.extractors.extractor_classes import MdaRecordingExtractor, MdaSortingExtractor
