+++ src/spikeinterface/sorters/external/tests/test_docker_containers.py
@@ -2,7 +2,7 @@ import os
 
 import pytest
 
-from spikeinterface import generate_ground_truth_recording
+from spikeinterface.generation import generate_ground_truth_recording
 from spikeinterface.core.core_tools import is_editable_mode
 import spikeinterface.sorters as ss
