+++ src/spikeinterface/sorters/tests/test_container_tools.py
@@ -5,7 +5,7 @@ import shutil
 import os
 
 import spikeinterface as si
-from spikeinterface import generate_ground_truth_recording
+from spikeinterface.generation import generate_ground_truth_recording
 
 from spikeinterface.sorters.container_tools import find_recording_folders, ContainerClient, install_package_in_container
 import platform
