+++ src/spikeinterface/sorters/tests/test_launcher.py
@@ -5,7 +5,7 @@ import time
 import pytest
 from pathlib import Path
 from platform import system
-from spikeinterface import generate_ground_truth_recording
+from spikeinterface.generation import generate_ground_truth_recording
 from spikeinterface.sorters import run_sorter_jobs, run_sorter_by_property
 
 # no need to have many
