+++ src/spikeinterface/sortingcomponents/motion/tests/test_motion_interpolation.py
@@ -4,7 +4,7 @@ import numpy as np
 import pickle
 
 from spikeinterface import NumpyRecording, load
-from spikeinterface.core import generate_ground_truth_recording
+from spikeinterface.generation import generate_ground_truth_recording
 from spikeinterface.core.motion import Motion
 from spikeinterface.core.testing import check_recordings_equal
 from spikeinterface.sortingcomponents.motion.motion_interpolation import (
@@ -241,7 +241,7 @@ def test_InterpolateMotionRecording():
             contact_shape_params={"radius": 6},
         ),
         generate_sorting_kwargs=dict(firing_rates=6.0, refractory_period_ms=4.0),
-        noise_kwargs=dict(noise_levels=5.0, strategy="on_the_fly"),
+        noise_kwargs=dict(noise_levels=5.0),
         seed=2205,
     )
