+++ src/spikeinterface/sortingcomponents/tests/common.py
@@ -1,4 +1,4 @@
-from spikeinterface.core import generate_ground_truth_recording
+from spikeinterface.generation import generate_ground_truth_recording
 
 
 def make_dataset():
@@ -16,7 +16,7 @@ def make_dataset():
             contact_shape_params={"radius": 6},
         ),
         generate_sorting_kwargs=dict(firing_rates=6.0, refractory_period_ms=4.0),
-        noise_kwargs=dict(noise_levels=5.0, strategy="on_the_fly"),
+        noise_kwargs=dict(noise_levels=5.0),
         seed=2205,
     )
