+++ src/spikeinterface/widgets/tests/test_widgets.py
@@ -13,9 +13,9 @@ if __name__ != "__main__":
 
 from spikeinterface import (
     compute_sparsity,
-    generate_ground_truth_recording,
     create_sorting_analyzer,
 )
+from spikeinterface.generation import generate_ground_truth_recording
 
 import spikeinterface.widgets as sw
 import spikeinterface.comparison as sc
@@ -54,7 +54,7 @@ class TestWidgets(unittest.TestCase):
                 contact_shape_params={"radius": 6},
             ),
             generate_sorting_kwargs=dict(firing_rates=10.0, refractory_period_ms=4.0),
-            noise_kwargs=dict(noise_levels=5.0, strategy="on_the_fly"),
+            noise_kwargs=dict(noise_levels=5.0),
             seed=2205,
         )
         # cls.recording = recording.save(folder=cache_folder / "recording")
